requirements.txt
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.1
fake-useragent==1.5.1
Flask==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
